Blog page and Markdown Parsing
Added a blog page, which pulls from a Strapi backend. Strapi is
a headless CMS-as-a-service provider. I have Strapi deployed locally
right now, and my blog page is set up as such. I will soon have to
refactor it and pull the fetch code out of that page and into a netlify
function, so that the API key won't be exposed.