Use api routes to proxy external request
Any functions placed in the /pages/api/ directory will be treated
as serverless functions, and will run only as a lambda once compiled.

This particular function will grab data from an external API for use
during the compilation step on Vercel. Note that the URL, and any API
keys that might be used, won't be exposed to a site user.