Do not return anything if auth check fails.
I am using the same "RequireAuth" component to check whether users
have access to an entire page AND for components on pages. If I return
something like "not authorized," then it will show up on all components
that I'm restricting access to. Like the delete button on the blog index
page. I don't want that.