Return Tiptapedit state to parent component
In this commit, I figured out how to return the state of the editor
to the parent by using a function that I pass in to the child from the
parent. This way, the state is always handled by the parent, but the
child can still send the data back.

This also mirrors my implementation of the <input> fields in the app,
which means that I can keep using my <form> onSubmit handler to save
things in my database!