Clear Tiptap body content on first focus
This was a fun one. I wanted a way to clear the content of the Tiptap
editor component's body on the first focus, and only the first focus.
Why? Because it seemed like fun.

Use a useEffect hook with an empty dependency array to set a stateful
variable that shows whether this editor has been focused before, and then
use that to conditionally clear it upon first focus.