Cache npm packages in CircleCI
Adds caching behavior to the CircleCI config. Saves a Cache
of the installed dependencies with a key linked to the particular version
of package-lock.json. If package.json has changed, it instead loads the
dependencies from the last CI run, and then runs `npm install` to update
any that have changed.

In all circumstances, it saves a cache after npm install but before the
test suite runs. This is to ensure that the cache is saved even if the
tests fail.