Implement changes to allow Cypress to run on CI
The latest version of Cypress, combined with the latest version
of Next.JS, introduced a regression in some production environments.
CircleCI's docker images were affected by this.

The regression can be found here:
https://github.com/vercel/next.js/issues/30750

The particular package there is different, but the cause and solution
were the same.

I downgraded Cypress to a previous version, and disabled the
"esmExternals" feature in Next.JS. This feature, which became standard
in Next v12, makes ECMA Modules the primary method of importing/exporting
code, over the older CommonJS module method.

Disabling this feature does not appear to affect the functionality of my
site in any way, despite my exclusive use of the new module methods.