Add CircleCI Workflow for Cypress & Coverage
Unlike Jest tests, Cypress tests run in a browser, and thus require
a built version of the application, along with a server to run it on.
This workflow does that.

The configuration is primarily composed of a Workflow, Jobs, and Steps.
Workflows have Jobs, and Jobs have steps.

You may also notice a section entitles "executors." This is simply a way
to store a variable, and I could also enter it directly into the job
configuration below. I could do the same for entire jobs.

Finally, there are orbs, which could be understood as external libraries.
I use the Node, Cypress, and Codecov orbs.

There are three jobs in this workflow. The first, node/test, runs my
jest-based unit tests. It also saves any artifacts from the tests, and
which should include code coverage reports, and uploads them to codecov.

The second job, cypress/install, sets up the environment that Cypress
will run the tests in later. It saves that "workspace" and then passes
it on to any later steps.

The final section, cypress/run, actually runs the test. I specify that
I want to use the chrome browser, that I want to use the available
workspace from the prior step, that I want to run the server before the
tests, and that I want to wait to start testing until the server responds
to a GET request. Like the jest tests, it also uploads the coverage results
to codecov.