Add Login page Cypress test
It is difficult to mock a social login with Cypress without using
real credentials. Even if I were to do that, it would likely fail in
CI due to anti-bot measures that the major login providers use.

As such, I am ignoring that area for now. I will change my login method
in a later PR to use a database that I control so that I can test parts
of the app that are behind authentication.