Remove Editor Route and Tests
I have spent the past few days trying to figure out how to protect my
"new page" route AND be able to test that authentication method with
Cypress. I've tried NextAuth and Cypress, and frankly, I'm sort of tired
of it now. I'll do this another time, but for now, I'm removing that route
and the tests that look at it, perhaps in favor of a managed CMS.