Update fetchAll to also return filename
Since I am using the filename as the identifier in my [posts] page's
getStaticProps and getStaticPaths, if I want to link one of those pages
I will need to use that as the URL.

Since I want to link to the posts from the index page, I need to pass
along the filenames to that page.

This refactors the fetchAll method to do that.