Adjust CircleCI config to support caching
Added an optional cache key to specify how the cache should be named.
Cypress was using a very restrictive naming scheme that resulted in
cache misses the majority of the time.

I have also adjusted the orb itself. I forked it and uploaded my own
version to CircleCI. In that version, I move the cache step to after
the build step, and also cache the `.next/cache` folder, which speeds
up future builds where that cache is valid.

I intend to bring much of the functionality of the orb into my own
CircleCI config so that I don't have to rely on a hacky personal version
for basic functionality.