Replace CircleCI orb jobs with custom job
In a previous commit, I reference work that I did on a fork of the
Cypress CircleCI orb to move the save_cache step to after the build
step. I do not feel comfortable using a hacky version of an official
orb, and would prefer to instead make my configuration entirely
custom.

This commit uses the Cypress orb as a starting point, and pulls out the
relevant branches into my own configuration. It simplifies places where
the orb had to include conditional logic or variables. Those are
necessary to support a wide variety of configurations, but not for my
own use.