Change sitemap generation to postbuild script
My original method for generating a sitemap used `getServerSideProps`,
which worked well locally, but did not work in production. This is
because production did not have a server with my markdown files to
get props from!

This overhauls the sitemap generation process by doing it at buildtime,
via a script that is intended to run after the build step.